fix: вынес функции из auth файла для использования в posts
