+++ server/auth.js
@@ -1,8 +1,8 @@
 import express, { json } from 'express';
 import { v4 as uuidv4 } from 'uuid';
 import bcrypt from 'bcrypt';
-import { generateTokens, tokenTimeParser } from '../src/common/common.js';
-import jwt from 'jsonwebtoken';
+import { generateTokens } from '../src/common/common.js';
+import { writeRefreshToBD, authMiddleware } from './auth_functions.js'
 import dotenv from 'dotenv';
 import cors from 'cors';
 import cookieParser from 'cookie-parser';
@@ -20,90 +20,6 @@ app.use(cors({
 app.use(json());
 app.use(cookieParser());
 
-
-//! Функция для записи Refresh-токена в БД ==========================>>>>>>>>
-async function writeRefreshToBD(userId, refresh) {
-  const refreshCreateQuery = `
-    INSERT INTO tokens (user_id, refresh_token, expires_at, created_at, is_active)
-    VALUES ($1, $2, $3, NOW(), true)
-    ON CONFLICT (user_id) 
-    DO UPDATE 
-    SET refresh_token = $2, 
-        expires_at = $3, 
-        created_at = NOW(), 
-        is_active = true;
-  `;
-  
-  const values = [userId, refresh, tokenTimeParser()];
-
-  try {
-    await queryDB(refreshCreateQuery, values);
-  } catch (error) {
-    console.error('Ошибка при записи refresh-токена в БД:', error);
-  }
-}
-
-//! Middleware для проверки Access-токена и автоматического обновления ==========================>>>>>>>>
-const authenticateToken = async (req, res, next) => {
-  const authHeader = req.headers.authorization;
-  if (!authHeader) return res.status(401).json({ message: 'Нет токена' });
-
-  const accessToken = authHeader.split(' ')[1];
-
-  jwt.verify(accessToken, process.env.JWT_SECRET, async (err, decoded) => {
-    if (!err) {
-      req.user = decoded;
-      req.tokens = {accessToken: accessToken, refreshToken: req.cookies.refreshToken}
-      return next();
-    }
-
-    if (err.name === 'TokenExpiredError') {
-      const refreshToken = req.cookies.refreshToken;
-      if (!refreshToken) return res.status(401).json({ message: 'Нет refresh-токена' });
-
-      try {
-        const result = await queryDB(
-          'SELECT user_id FROM tokens WHERE refresh_token = $1 AND is_active = true',
-          [refreshToken]
-        );
-
-        if (result.data.length === 0) return res.status(403).json({ message: 'Refresh-токен не найден' });
-
-        jwt.verify(refreshToken, process.env.JWT_SECRET, async (err, decodedRT) => {
-          if (err) return res.status(403).json({ message: 'Refresh-токен недействителен' });
-          
-          const newTokens = generateTokens({id: decodedRT.id, email: decodedRT.email})
-
-          res.setHeader('Authorization', `Bearer ${newTokens.accessToken}`); // Отправляем новый AT
-
-          await writeRefreshToBD(decodedRT.id, newTokens.refreshToken);
-
-          res.cookie('refreshToken', newTokens.refreshToken, {
-            httpOnly: true,
-            sameSite: 'Strict',
-            maxAge: 24 * 60 * 60 * 1000,
-          });
-          req.user = decodedRT;
-          next();
-        });
-      } catch (error) {
-        return res.status(500).json({ message: 'Ошибка сервера' });
-      }
-    } else {
-      return res.status(401).json({ message: 'Токен недействителен' });
-    }
-  });
-};
-
-//? Функция для проверки доступа по токенам ==========================>>>>>>>>
-export const authMiddleware = async (req, res, next) => {
-  try {
-    await authenticateToken(req, res, next);
-  } catch (error) {
-    console.error('Ошибка проверки токена:', error.message);
-    res.status(403).json({ message: 'Неавторизован' });
-  }
-};
 
 //! Регистрация пользователя ==========================>>>>>>>>
 app.post('/add-user', async (req, res) => {
@@ -211,5 +127,5 @@ app.get('/get-user', authMiddleware, async (req, res) => {
 
 //! Запуск сервера ==========================>>>>>>>>
 app.listen(port, () => {
-  console.log(`Сервер работает на ${port}`);
+  console.log(`Сервер auth работает на ${port}`);
 });
